Initial commit after relocation
citometro files have been relocated to projectos/cad/py_freecad/citometro
This has been done to also relocate its dependences and be able to
develop them independently to the citometro project.
The reason is that its dependences are also used in other projects, such as
the corexy 3d printer.
This first commit just have some of the base aluminum extrusion and shaft
holder. So it is just a very basic first version