Add Python FreeCad module subtree in modules/comps/
Creation of the git subtree to have the functions and classes independent
from citometro. This module will be used by other scripts such as the
CoreXY 3D printer, and any other new scripted FreeCad design I may do.
Having the module in a subtree allows to have independent development
but also keep track of the version of the module that have been used in
a particular citometro version